moved ModelManipulator
